+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -3,9 +3,9 @@ package kitchenpos.products.application;
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.domain.Product;
-import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.infra.PurgomalumClient;
+import kitchenpos.products.tobe.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,46 +33,41 @@ public class ProductService {
 
     @Transactional
     public Product create(final Product request) {
-        final BigDecimal price = request.getPrice();
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+        String name = request.getName().getName();
+        if (purgomalumClient.containsProfanity(name)) {
             throw new IllegalArgumentException();
         }
-        final String name = request.getName();
-        if (Objects.isNull(name) || purgomalumClient.containsProfanity(name)) {
-            throw new IllegalArgumentException();
-        }
-        final Product product = new Product();
-        product.setId(UUID.randomUUID());
-        product.setName(name);
-        product.setPrice(price);
+        Product product = new Product(request.getName(), request.getPrice());
         return productRepository.save(product);
     }
 
-    @Transactional
-    public Product changePrice(final UUID productId, final Product request) {
-        final BigDecimal price = request.getPrice();
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-            throw new IllegalArgumentException();
-        }
-        final Product product = productRepository.findById(productId)
-            .orElseThrow(NoSuchElementException::new);
-        product.setPrice(price);
-        final List<Menu> menus = menuRepository.findAllByProductId(productId);
-        for (final Menu menu : menus) {
-            BigDecimal sum = BigDecimal.ZERO;
-            for (final MenuProduct menuProduct : menu.getMenuProducts()) {
-                sum = sum.add(
-                    menuProduct.getProduct()
-                        .getPrice()
-                        .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
-                );
-            }
-            if (menu.getPrice().compareTo(sum) > 0) {
-                menu.setDisplayed(false);
-            }
-        }
-        return product;
-    }
+//    @Transactional
+//    public Product changePrice(final UUID productId, final Product request) {
+//        Product product1 = new Product(productId, request.getName(), request.getPrice());
+//
+//        final BigDecimal price = request.getPrice();
+//        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+//            throw new IllegalArgumentException();
+//        }
+//        final Product product = productRepository.findById(productId)
+//            .orElseThrow(NoSuchElementException::new);
+//        product.setPrice(price);
+//        final List<Menu> menus = menuRepository.findAllByProductId(productId);
+//        for (final Menu menu : menus) {
+//            BigDecimal sum = BigDecimal.ZERO;
+//            for (final MenuProduct menuProduct : menu.getMenuProducts()) {
+//                sum = sum.add(
+//                    menuProduct.getProduct()
+//                        .getPrice()
+//                        .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
+//                );
+//            }
+//            if (menu.getPrice().compareTo(sum) > 0) {
+//                menu.setDisplayed(false);
+//            }
+//        }
+//        return product;
+//    }
 
     @Transactional(readOnly = true)
     public List<Product> findAll() {
